modularized get_plant_biomes
