+++ actions/__init__.py
@@ -3,3 +3,4 @@ from .release_animal import release_animal
 from .report import build_facility_report
 from .feed_animal import feed_animal
 from .cultivate_plant import cultivate_plant
+from .get_plant_biomes import get_plant_biomes
